requirements.txt
certifi==2018.11.29
Django==1.11.17
Pillow==5.4.1
pytz==2018.9
selenium==3.141.0
urllib3==1.24.1
wincertstore==0.2
